Phase 2: Refactor Sources Actions & Page
